# frontend/docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Orient-advertise-backend API",
    "description" : "Multi-module Spring Boot platform for TV-box content delivery — device management, content uploads + transcoding, playlists, scheduling, remote actions, incident lifecycle, proof-of-play, aggregated reporting, Excel export, and external partner integrations.\n\n**Authentication.** Most endpoints require a JWT access token (`Authorization: Bearer <token>`) obtained from `POST /auth/login`. External integration endpoints under `/service/external/**` use an `X-API-Key` header instead — see the *External* group.\n\n**Sensitive endpoints.** Operations marked with `[SENSITIVE]` in their summary affect production state, expose user data, or irreversibly delete records. Audit logging is enabled and the operator's username is recorded.\n",
    "contact" : {
      "name" : "orientadvertise.services"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version" : "1.0.55"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Current host"
  }, {
    "url" : "https://api.example.com",
    "description" : "Production"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/devices/{id}/remote" : {
      "get" : {
        "tags" : [ "Remote Control" ],
        "summary" : "Current remote session for a device",
        "operationId" : "current",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The live session (no viewerTicket)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No live session for this device",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Caller is VIEWER/ADVERTISER",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Remote Control", "Sensitive" ],
        "summary" : "[SENSITIVE] Start a remote view/control session",
        "operationId" : "start",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StartRemoteSessionRequest"
              }
            }
          }
        },
        "responses" : {
          "404" : {
            "description" : "Device not found, soft-deleted, or out of scope",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          },
          "503" : {
            "description" : "Remote control is disabled on this server",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          },
          "201" : {
            "description" : "Session created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "Device reported capability.supported == false",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Caller is VIEWER/ADVERTISER",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "A PENDING/ACTIVE session already exists",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RemoteSessionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/refresh" : {
      "post" : {
        "tags" : [ "Auth" ],
        "summary" : "Rotate the refresh cookie and mint a fresh access token",
        "description" : "Single-use refresh: the supplied cookie value is invalidated and replaced. Reuse of a previously rotated id is treated as a possible replay and invalidates the entire token family (every refresh token issued under the same login session). Role changes take effect here — the new access token is minted with the user's CURRENT role. The refresh token is read from the `refresh_token` cookie; a fresh cookie is set on the response.\n",
        "operationId" : "refresh",
        "parameters" : [ {
          "name" : "refresh_token",
          "in" : "cookie",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "401" : {
            "description" : "Cookie missing, token unknown, expired, or replay detected",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "New access token in body, rotated refresh token in cookie",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccessTokenResponse"
                }
              }
            }
          }
        },
        "security" : [ ]
      }
    },
    "/api/auth/logout" : {
      "post" : {
        "tags" : [ "Auth" ],
        "summary" : "Invalidate a refresh-token family (logout)",
        "description" : "Reads the `refresh_token` cookie, invalidates every refresh token in its family, and clears the cookie on the response. Returns 204 even when the cookie is missing or the token is unknown — no information leak.",
        "operationId" : "logout",
        "parameters" : [ {
          "name" : "refresh_token",
          "in" : "cookie",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Logged out (or token already invalid)"
          }
        },
        "security" : [ ]
      }
    },
    "/api/auth/login" : {
      "post" : {
        "tags" : [ "Auth" ],
        "summary" : "Authenticate and receive an access token",
        "description" : "Validates credentials, then returns a short-lived JWT access token in the response body and issues a long-lived refresh token as an HttpOnly cookie (`refresh_token`, scoped to `/api/auth`, `SameSite` configurable, default `Strict`). The refresh token is never exposed to JavaScript. Reject reasons: invalid credentials (401), deactivated account (401), no role assigned (401).\n",
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              },
              "examples" : {
                "credentials" : {
                  "description" : "credentials",
                  "value" : {
                    "username" : "alice",
                    "password" : "s3cret-passphrase"
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Authenticated; access token in body, refresh token in `refresh_token` cookie",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccessTokenResponse"
                },
                "examples" : {
                  "ok" : {
                    "description" : "ok",
                    "value" : {
                      "accessToken" : "eyJhbGciOiJIUzI1NiJ9.eyJzdWIi..."
                    }
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Invalid credentials, deactivated account, or no role assigned",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "StartRemoteSessionRequest" : {
        "type" : "object",
        "properties" : {
          "viewOnly" : {
            "type" : "boolean"
          }
        }
      },
      "CapabilityDto" : {
        "type" : "object",
        "properties" : {
          "supported" : {
            "type" : "boolean"
          },
          "input" : {
            "type" : "string"
          },
          "transport" : {
            "type" : "string"
          },
          "maxWidth" : {
            "type" : "integer",
            "format" : "int32"
          },
          "maxHeight" : {
            "type" : "integer",
            "format" : "int32"
          },
          "reportedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "RemoteSessionResponse" : {
        "type" : "object",
        "properties" : {
          "sessionId" : {
            "type" : "string"
          },
          "deviceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "status" : {
            "type" : "string"
          },
          "relayUrl" : {
            "type" : "string"
          },
          "viewerTicket" : {
            "type" : "string"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "viewOnly" : {
            "type" : "boolean"
          },
          "deliveredVia" : {
            "type" : "string"
          },
          "capability" : {
            "$ref" : "#/components/schemas/CapabilityDto"
          }
        }
      },
...
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "correlationId" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string"
          },
          "fieldErrors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FieldError"
            }
          },
          "details" : {
            "type" : "object"
          }
        }
      },
      "FieldError" : {
        "type" : "object",
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "rejectedValue" : {
            "type" : "object"
          }
        }
      },
      "AccessTokenResponse" : {
        "type" : "object",
        "properties" : {
          "accessToken" : {
            "type" : "string",
            "description" : "Short-lived JWT access token. Send as `Authorization: Bearer <token>`.",
            "example" : "eyJhbGciOiJIUzI1NiJ9.eyJzdWIi..."
          }
        },
        "description" : "Access-token response body. The refresh token is delivered separately as an HttpOnly cookie."
      },
      "LoginRequest" : {
        "required" : [ "password", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "password" : {
            "maxLength" : 2147483647,
            "minLength" : 4,
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "JWT access token from `POST /auth/login`. Roles encoded in the token: ADMIN, OPERATOR, VIEWER, ADVERTISER.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "apiKeyAuth" : {
        "type" : "apiKey",
        "description" : "External-integration API key. Issued by ADMIN via `POST /service/admin/service-keys`. Rate-limited 100 req/hr per key. Revocation is immediate — no in-memory cache widens the window.",
        "name" : "X-API-Key",
        "in" : "header"
      }
    }
  }
}